test(#1588): fold init regression into init suite
